refactor: move oauth callback forwarder
